HSEARCH-4703 Add Lucene specific module path based tests
