+++ integrationtest/jdk/java-modules/orm-lucene/src/main/resources/hibernate.properties
@@ -1,0 +1,16 @@
+# Hibernate ORM properties:
+## Connection info: see integration-test parent POM
+hibernate.hbm2ddl.auto = create-drop
+hibernate.show_sql = true
+hibernate.format_sql = true
+hibernate.max_fetch_depth = 5
+# We can't use classes from the hibernate-testing module unless we add an explicit dependency to that module.
+#hibernate.cache.region_prefix hibernate.test
+#hibernate.cache.region.factory_class = org.hibernate.testing.cache.CachingRegionFactory
+
+# Hibernate Search properties:
+## Connection info: see POM
+hibernate.search.backend.type=lucene
+hibernate.search.automatic_indexing.synchronization.strategy = sync
+hibernate.search.backend.directory.type = local-heap
+hibernate.search.backend.analysis.configurer = org.hibernate.search.integrationtest.java.modules.orm.lucene.config.MyLuceneAnalysisConfigurer
