--- integrationtest/jdk/java-modules/src/main/java/org/hibernate/search/integrationtest/java/module/entity/Author.java
@@ -1,42 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.integrationtest.java.module.entity;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-import org.hibernate.search.integrationtest.java.module.config.MyElasticsearchAnalysisConfigurer;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.FullTextField;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
-
-@Entity
-@Indexed
-public class Author {
-
-	@Id
-	@GeneratedValue
-	private Integer id;
-
-	@FullTextField(analyzer = MyElasticsearchAnalysisConfigurer.MY_ANALYZER)
-	private String name;
-
-	public Author() {
-	}
-
-	public Integer getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-}
